+++ processes/addFulfillManifest.py
@@ -46,21 +46,37 @@ class FulfillProcess(CoreProcess):
         '''Load batch of LA works based on startTimeStamp'''
 
         batches = self.load_batches(self.prefix, self.s3Bucket)
-        filtered_batches = batches.search(f"Contents[?to_string(LastModified) > '\"{startTimeStamp}\"'].Key")
-        for batch in filtered_batches:
-            currKey = batch['Key']
-            metadataObject = self.getObjectFromBucket(currKey, self.s3Bucket)
-            self.update_manifest(metadataObject, self.s3Bucket, currKey)
+        if startTimeStamp:
+            filtered_batches = batches.search(f"Contents[?to_string(LastModified) > '\"{startTimeStamp}\"'].Key")
+            for batch in filtered_batches:
+                for c in batch['Contents']:
+                    currKey = c['Key']
+                    metadataObject = self.getObjectFromBucket(currKey, self.s3Bucket)
+                    self.update_manifest(metadataObject, self.s3Bucket, currKey)
+        else:
+            for batch in batches:
+                for c in batch['Contents']:
+                    currKey = c['Key']
+                    metadataObject = self.getObjectFromBucket(currKey, self.s3Bucket)
+                    self.update_manifest(metadataObject, self.s3Bucket, currKey)
 
     def update_manifest(self, metadataObject, bucketName, currKey):
 
         metadataJSON = json.loads(metadataObject['Body'].read().decode("utf-8"))
         metadataJSONCopy = copy.deepcopy(metadataJSON)
         
-        metadataJSON = self.linkFulfill(metadataJSON)
-        metadataJSON = self.readingOrderFulfill(metadataJSON)
-        metadataJSON = self.resourceFulfill(metadataJSON)
-        metadataJSON = self.tocFulfill(metadataJSON)
+        counter = 0
+    
+        metadataJSON, counter = self.linkFulfill(metadataJSON, counter)
+        metadataJSON, counter = self.readingOrderFulfill(metadataJSON, counter)
+        metadataJSON, counter = self.resourceFulfill(metadataJSON, counter)
+        metadataJSON, counter = self.tocFulfill(metadataJSON, counter)
+
+        if counter >= 4: 
+            for link in metadataJSON['links']:
+                self.fulfillFlagUpdate(link)
+
+        self.closeConnection()
 
         if metadataJSON != metadataJSONCopy:
             try:
@@ -70,31 +86,32 @@ class FulfillProcess(CoreProcess):
                 )
             except ClientError as e:
                 logging.error(e)
-        
-        self.closeConnection()
 
     def linkFulfill(self, metadataJSON):
         for link in metadataJSON['links']:
-            link['href'] = self.fulfillReplace(self, link)
+            fulfillLink, counter = self.fulfillReplace(link, counter)
+            link['href'] = fulfillLink
 
-        return metadataJSON
+        return (metadataJSON, counter)
             
     def readingOrderFulfill(self, metadataJSON):
         for readOrder in metadataJSON['readingOrder']:
-            readOrder['href'] = self.fulfillReplace(self, readOrder)
+            fulfillLink, counter = self.fulfillReplace(readOrder, counter)
+            readOrder['href'] = fulfillLink
 
-        return metadataJSON
+        return (metadataJSON, counter)
 
     def resourceFulfill(self, metadataJSON):
         for resource in metadataJSON['resources']:
-            resource['href'] = self.fulfillReplace(self, resource)
+            fulfillLink, counter = self.fulfillReplace(resource, counter)
+            resource['href'] = fulfillLink
 
-        return metadataJSON
+        return (metadataJSON, counter)
 
     def tocFulfill(self, metadataJSON): 
 
         '''
-        THe toc dictionary has no "type" key like the previous dictionaries 
+        The toc dictionary has no "type" key like the previous dictionaries 
         therefore the 'href' key is evaluated instead
         '''
 
@@ -103,18 +120,33 @@ class FulfillProcess(CoreProcess):
                 or 'epub' in toc['href']:
                     for link in self.session.query(Link) \
                         .filter(Link.url == toc['href'].replace('https://', '')):
-                            toc['href'] = f'{self.host}:{self.port}/fulfill/{link.id}'
-                            link.flags['fulfill_limited_access'] = True
-        return metadataJSON
+                            counter += 1
+                            toc['href'] = f'http://{self.host}:{self.port}/fulfill/{link.id}'
+
+        return (metadataJSON, counter)
 
     def fulfillReplace(self, metadata):
         if metadata['type'] == 'application/pdf' or metadata['type'] == 'application/epub+zip' \
             or metadata['type'] == 'application/epub+xml':
                 for link in self.session.query(Link) \
                     .filter(Link.url == metadata['href'].replace('https://', '')):
-                        metadata['href'] = f'{self.host}:{self.port}/fulfill/{link.id}'
-                        link.flags['fulfill_limited_access'] = True
-        return metadata['href']
+                        counter += 1            
+                        metadata['href'] = f'http://{self.host}:{self.port}/fulfill/{link.id}'
+
+        return (metadata['href'], counter)
+    
+    def fulfillFlagUpdate(self, metadata):
+        if metadata['type'] == 'application/webpub+json':
+            for link in self.session.query(Link) \
+                .filter(Link.url == metadata['href'].replace('https://', '')):   
+                        print(link)
+                        print(link.flags)
+                        if 'fulfill_limited_access' in link.flags.keys():
+                            if link.flags['fulfill_limited_access'] == False:
+                                newLinkFlag = dict(link.flags)
+                                newLinkFlag['fulfill_limited_access'] = True
+                                link.flags = newLinkFlag
+                                self.commitChanges()
 
 class FulfillError(Exception):
     pass